lesson-14/requirements.txt
# 사용자 맞춤형 자동매매 시스템 의존성
# Python 3.13 호환 버전

# 웹 프레임워크
Flask>=3.0.0
Flask-CORS>=4.0.0

# 데이터 처리  
pandas>=2.1.0
numpy>=1.26.0

# 데이터베이스
SQLAlchemy>=2.0.23

# 유틸리티
PyYAML>=6.0.1
python-dotenv>=1.0.0

# 테스트
pytest>=7.4.0
pytest-cov>=4.1.0

# 타입 체크
mypy>=1.7.0

# 코드 품질
black>=23.11.0
flake8>=6.1.0

